- no renewal deposit ends when maturated - renewal no capitalization deposit deposits the interest in the account and restarts - renewal cappitalization deposits accumulates the interest and restarts - checkDepositView shows in the center of the screen - account has now float instead of int as money type; as well as the deposit

+++ src/main/java/itsix/bank_deposit/logic/IAccount.java
@@ -1,32 +1,32 @@
 package itsix.bank_deposit.logic;
 
-import java.util.List;
-
 import itsix.bank_deposit.exception.InvalidOperationException;
 import itsix.bank_deposit.publisher_subscriber.IPublisher;
 
+import java.util.List;
+
 public interface IAccount extends IPublisher {
 
-	String getCurrencyName();
+    String getCurrencyName();
 
-	String getCurrencySymbol();
+    String getCurrencySymbol();
 
-	float getBalance();
+    float getBalance();
 
-	float getCreditBalance();
+    float getCreditBalance();
 
-	float getDebitBalance();
+    float getDebitBalance();
 
-	ICurrency getCurrency();
+    ICurrency getCurrency();
 
-	void deposit(int money);
+    void deposit(float money);
 
-	void withdraw(int money) throws InvalidOperationException;
+    void withdraw(float money) throws InvalidOperationException;
 
-	List<ICurrency> subtractOwnCurrency(List<ICurrency> remainingCurrencies);
+    List<ICurrency> subtractOwnCurrency(List<ICurrency> remainingCurrencies);
 
-	boolean hasCurrency(ICurrency currency);
+    boolean hasCurrency(ICurrency currency);
 
-	boolean hasFunds(int money);
+    boolean hasFunds(float money);
 
 }
