+++ src/main/java/itsix/bank_deposit/logic/IInterestCalculator.java
@@ -2,7 +2,7 @@ package itsix.bank_deposit.logic;
 
 public interface IInterestCalculator {
 
-	float calculateDailyIncome(int money);
+	float calculateDailyIncome(float money);
 
 	float getInterest();
 }
